pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "vek-parser"
version = "0.1.0"
description = "Универсальный парсер для сбора структурированных данных с веб-сайтов"
readme = "README.md"
authors = [
    { name = "Emil Markov", email = "markovemil2@gmail.com" }
]
license = "MIT"
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
]
keywords = ["web", "parser", "scraping", "crawler"]
dependencies = [
    "requests>=2.25.0",
    "selectorlib>=0.16.0",
    "parsel>=1.6.0",
    "PyYAML>=5.4.0",
]
requires-python = ">=3.9"

[project.urls]
Homepage = "https://github.com/EmilMarkov/VekParser"
Repository = "https://github.com/EmilMarkov/VekParser.git"
Documentation = "https://github.com/EmilMarkov/VekParser#readme"

[tool.hatch.build.targets.wheel]
packages = ["src/vek_parser"]
